requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.5
flask-migrate==2.3.0
flask-bcrypt
pyjwt
flask-cors
